Use rb_eArgError directly.
It's part of the public API, so little point keeping another
private reference to it.

Additionally when storing an object in a global variable you MUST
register that variable to the gc with `rb_global_variable(&cArgumentError)`
and that *before* you store anything in it, so that the GC knows to
update your global variable if it moves that object.

In practice `ArgumentError` is pinned for now, so it would have been
OK, but it might not be pinned forever.